refactor: cleanup allowed option types (#3651)
* refactor: cleanup allowed option types

* formatting

* remove includes

* Fix tests

* fix usage

* keep double for save resume

* fmt

* remove double option type

* remove double option type

* Update python/pylibvw.cc

+++ vowpalwabbit/ftrl.cc
@@ -498,7 +498,7 @@ base_learner* ftrl_setup(VW::setup_base_i& stack_builder)
   if (!all.holdout_set_off)
   {
     all.sd->holdout_best_loss = FLT_MAX;
-    b->early_stop_thres = options.get_typed_option<size_t>("early_terminate").value();
+    b->early_stop_thres = options.get_typed_option<uint64_t>("early_terminate").value();
   }
 
   auto predict_ptr = (all.audit || all.hash_inv) ? predict<true> : predict<false>;
